Refactor Repository Folder
