+++ Capstone_API/Capstone_API/Controllers/SignUpController.cs
@@ -1,5 +1,5 @@
-﻿using Capstone_API.IRepository;
-using Capstone_API.Models;
+﻿using Capstone_API.Models;
+using Capstone_API.Repository.Interface;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
