+++ Capstone_API/Capstone_API/Repository/Interface/ILoginRepository.cs
@@ -2,11 +2,11 @@
 using System.Collections.Generic;
 using System.Xml.Serialization;
 
-namespace Capstone_API.IRepository
+namespace Capstone_API.Repository.Interface
 {
-    public interface ILoginRepository 
+    public interface ILoginRepository
     {
-        
+
         public Account GetAccount(string phone, string password);
     }
 }
